+++ packages/keyring-controller/src/KeyringController.ts
@@ -2194,7 +2194,7 @@ export class KeyringController extends BaseController<
       // NOTE: Not all keyrings implement this method in a asynchronous-way. Using `await` for
       // non-thenable will still be valid (despite not being really useful). It allows us to cover both
       // cases and allow retro-compatibility too.
-      await keyring.generateRandomMnemonic();
+      keyring.generateRandomMnemonic();
       await keyring.addAccounts(1);
     }