+++ packages/profile-sync-controller/src/shared/env.test.ts
@@ -1,5 +1,5 @@
-import type { MockVariable } from '../sdk/__fixtures__/test-utils';
 import { getEnvUrls, Env, Platform, getOidcClientId } from './env';
+import type { MockVariable } from '../sdk/__fixtures__/test-utils';
 
 describe('getEnvUrls', () => {
   it('should return URLs if given a valid environment', () => {