+++ packages/transaction-controller/src/TransactionController.ts
@@ -3298,10 +3298,12 @@ export class TransactionController extends BaseController<
     provider,
     blockTracker,
     chainId,
+    networkClientId,
   }: {
     provider: Provider;
     blockTracker: BlockTracker;
     chainId: Hex;
+    networkClientId: NetworkClientId;
   }): PendingTransactionTracker {
     const ethQuery = new EthQuery(provider);
 
@@ -3309,6 +3311,7 @@ export class TransactionController extends BaseController<
       blockTracker,
       getChainId: () => chainId,
       getEthQuery: () => ethQuery,
+      getNetworkClientId: () => networkClientId,
       getTransactions: () => this.state.transactions,
       isResubmitEnabled: this.#pendingTransactionOptions.isResubmitEnabled,
       getGlobalLock: () =>